refactor(core): export reference link APIs publicly
