+++ packages/core/package.json
@@ -13,7 +13,6 @@
   ],
   "exports": {
     ".": "./src/index.ts",
-    "./internal": "./src/internal.ts",
     "./style.css": "./src/style.css"
   },
   "files": [
@@ -64,7 +63,6 @@
   "publishConfig": {
     "exports": {
       ".": "./dist/index.js",
-      "./internal": "./dist/internal.js",
       "./style.css": "./dist/style.css"
     },
     "scripts": {},
