--- packages/core/src/internal.ts
@@ -1,10 +1,0 @@
-export {
-  inlineTextToMarkChunksWithContext,
-  type InlineMarkContext,
-} from './extensions/inline-text-to-mark-chunks.ts'
-export {
-  collectReferenceDefinitions,
-  type ReferenceDefinition,
-  type ReferenceDefinitionIndex,
-  type ReferenceDefinitions,
-} from './extensions/reference-links.ts'
