+++ packages/core/tsdown.config.ts
@@ -1,7 +1,7 @@
 import { defineConfig } from 'tsdown'
 
 export default defineConfig({
-  entry: ['src/index.ts', 'src/internal.ts', 'src/style.css'],
+  entry: ['src/index.ts', 'src/style.css'],
   platform: 'neutral',
   sourcemap: 'hidden',
   minify: true,
